Change type of code_length

+++ config.h
@@ -1,7 +1,9 @@
+#include <stddef.h>
+
 struct lipi_config {
 	unsigned long period;
 	unsigned long threshold;
-	unsigned long code_length;
+	size_t code_length;
 	char start_code;
 	char end_code;
 };
